1.merge forked source from origin manully 2.move test files in different folder
