+++ src/main/resources/application.properties
@@ -13,7 +13,7 @@ spring.datasource.password=NewPassword
 spring.datasource.hikari.minimum-idle=5
 spring.datasource.hikari.maximum-pool-size=15
 spring.datasource.hikari.auto-commit=true
-spring.datasource.hikari.idle-timeout=30000
+spring.datasource.hikari.idle-timeout=10000
 spring.datasource.hikari.pool-name=hikariCP
 spring.datasource.hikari.max-lifetime=30000
 spring.datasource.hikari.connection-timeout=30000
