+++ test/ltd/newbee/mall/test/TestCompareWithJunit.java
@@ -1,4 +1,4 @@
-package ltd.newbee.mall.controller.mall;
+package ltd.newbee.mall.test;
 
 import static org.junit.Assert.assertEquals;
