+++ test/ltd/newbee/mall/test/TestStudentController.java
@@ -1,4 +1,4 @@
-package ltd.newbee.mall.controller.mall;
+package ltd.newbee.mall.test;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.samePropertyValuesAs;
